Version 4

+++ PrestaZon/PrestaZon-ejb/src/java/metier/GestionUtilisateurLocal.java
@@ -6,6 +6,7 @@
 package metier;
 
 import entitees.Utilisateur;
+import java.math.BigDecimal;
 import java.util.List;
 import javax.ejb.Local;
 
@@ -16,13 +17,13 @@ import javax.ejb.Local;
 @Local
 public interface GestionUtilisateurLocal {
         
-    long ajoutUtilisateur(String nom, String prenom, String motDePasse, String adresse, long telephone, String mail, int idCompte) throws exceptions.UtilisateurExistantException;
+    BigDecimal ajoutUtilisateur(String nom, String prenom, String motDePasse, String adresse, long telephone, String mail, int idCompte) throws exceptions.UtilisateurExistantException;
     
     //long chercherUtilisateur(Integer id) throws exceptions.UtilisateurInconnuException;
 
     List<Utilisateur> listeUtilisateurs() throws exceptions.UtilisateurInconnuException;
 
-    void supprimer(long idUtilisateur) throws exceptions.UtilisateurInconnuException;
+    void supprimer(BigDecimal idUtilisateur) throws exceptions.UtilisateurInconnuException;
     
     //liste de tout les utilisateur
